Extract common Lua codes
