+++ src/lib.rs
@@ -2,5 +2,6 @@ pub mod config;
 pub mod filewatcher;
 pub mod handler;
 pub mod leader_election;
+pub mod lua;
 pub mod reconcile;
 pub mod types;
